First pass of restructure, moved Python files together into one package and added support files for automated testing
