+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+# Contributing to PyGmod
+
+## Python Extension & Package
+
+The Python extension and packages are contained in ``/pygmod/``. A Docker container has been provided for running tests:
+
+```
+docker build -f .Dockerfile .
+```
