--- bin_modules/main/python_extensions/luapyobject.hpp
@@ -1,8 +1,0 @@
-// Lua metatable for representing Python objects.
-
-#pragma once
-#include <GarrysMod/Lua/Interface.h>
-
-using namespace GarrysMod::Lua;
-
-void createLuaPyObjectMetatable(ILuaBase* lua);
