+++ docker-compose.yml
@@ -1,0 +1,6 @@
+version: '3'
+services:
+  pygmod_package_tests:
+    build:
+      context: ./pygmod/
+      dockerfile: Dockerfile
