+++ pygmod/pygmod/_streams.py
