+++ pygmod/pygmod/error_notif.py
