+++ pygmod/python_extensions/stack_utils.cpp
