+++ pygmod/setup.cfg
@@ -1,0 +1,12 @@
+[aliases]
+test=pytest
+
+[tool:pytest]
+addopts = --verbose --pylint --cov=gmod --cov-config=setup.cfg
+
+[coverage:run]
+source = gmod
+
+[coverage:report]
+show_missing = True
+precision = 2
