+++ travis.yml
@@ -1,0 +1,5 @@
+services:
+  - docker
+
+script:
+  - docker-compose up --build
